desktop: adding a sentence to the AppStream metadata.
I've been hesitating about this one, because it breaks string freeze (sorry,
translators!). But that's quite a major backport, so I'm adding it as a separate
paragraph element so that it won't break translation of previous sentence at
least. In worst case, the first sentence will be updated but not the second.