README: Port to Markdown
For developers, this format is a bit easier to read than the plain text
due to syntax mark on IDEs. For first time contributors, it is easy to
read, for example, on smarphone: responsiveness is taken into account by
the "HTML" preview. Also, the HTML preview makes easier to click on links.

These MD features above are objective, undisputed and unrefutable.