Issue #11385: revert creating/deleting a config dir for every in-build GIMP calls.
While theoretically, commit 2f9881c03fd was better (perfectly fresh
config directory then cleanup at the end of each run of GIMP during the
build), it looks that in practice, it makes the build a lot slower, in
particular on Windows and macOS that have more assets to build.

So I'm reverting to how it was before, except that I add a unique
cleanup step (at the start of the build only, just before creating a new
directory) and also, I get slightly better "depends" organization.