gitlab-ci, build: Make "Cross" scripts compatible with Local builds (part 1)
* Some bashisms were dropped, since local machines can have different shell
* Every env (sh and "cross") command is self-contained in only one script per
  stage (*deps*cross* and *gimp*cross* .sh), as we do in native scripts.

Also, reintroduce pkg caching for crossroad. The possibility of using it is
one in thousands since we don't have distributed cache, but it's better than
rely totally on MSYS2 mirrors, which can be be extremely sick sometimes.

Unfortunately, the final result of my cleanup and compatibility work is limited,
since our (not anymore) "cross" builds depends totally on a Debian build.