plug-ins: fix blank lines in histogram-export (on Windows)
Resolves #6096
On Windows, csv.writer() will add an additional newline if
the file is opened in 'wt' mode. Per the documentation, it
should be opened in 'wb' mode to avoid this - however, that
can have an impact on the output in Linux systems.
This patch uses sys.platform to check if we're on a Windows
system, and if so, we use 'wb'. Otherwise, we open in 'wt' mode.
Note that this solution differs from the Python 3 fix implemented
in 7f9e819b for GIMP 3.0.