widgets: Prevent locks from being removed when a mask is added
Currently, masks are set to editable when
clicked in the layer tree view. This may
not be desirable if the layer is pixel or
position locked, as those locks will be
automatically removed.
This patch adds a check to ensure the
layer is actually editable before setting
it to be so when clicked.