plug-ins: Fix crash in non-interactive mode for file-tga/xpm
Fixes issue #11839 for file-tga.c and file-xpm.c
These two plug-ins always call gimp_ui_init () when exporting an
image, even in non-interactive mode. This can cause problems
when the plugins are run via scripts.
This patch moves the line to after a check for 
run_mode == INTERACTIVE is performed, to ensure that we're running
the plug-ins from the GUI instead of a script.