images: generate background image for installation on macOS.
Follow up of #5507, the background image where one drops GIMP logo into
applications also contains GIMP logo.

I generate the new image myself using the existing background
(originally in the gimp-macos-build repo in package/gimp-dmg.xcf, with
the old logo now removed) and the new logo, exactly in the middle
position (horizontally).

Eventually the whole background should likely be redesigned to something
fancier, ideally in SVG, and the logo position should possibly be
hand-picked (our logo is not symmetrical and I remember that Aryeom was
saying it should not always be centered to feel centered). But for now,
that will do.